@@ -9,7 +9,7 @@
         "composer"
     ],
     "homepage": "https://github.com/composer/satis",
-    "type": "project",
+    "type": "composer-plugin",
     "license": "MIT",
     "authors": [
         {
@@ -39,14 +39,19 @@
     ],
     "replace": {
         "symfony/polyfill-php72": "*",
-        "symfony/polyfill-php73": "*"
+        "symfony/polyfill-php73": "*",
+        "symfony/polyfill-php74": "*",
+        "symfony/polyfill-php80": "*",
+        "symfony/polyfill-php81": "*"
     },
     "require": {
-        "php": "^7.4||^8.0",
-        "composer/composer": "^2.1.9",
-        "symfony/console": "^4.4||^5.0",
-        "twig/twig": "^3.3",
-        "twig/html-extra": "^3.3"
+        "php": "^8.1",
+        "ext-zip": "*",
+        "composer/composer": "^2.6.4",
+        "composer-plugin-api": "^2.3.0",
+        "symfony/console": "^6.3.4",
+        "twig/twig": "^3.7.1",
+        "twig/html-extra": "^3.7.1"
     },
     "autoload": {
         "psr-4": {
@@ -55,8 +60,8 @@
     },
     "require-dev": {
         "ext-json": "*",
-        "mikey179/vfsstream": "^1.6.8",
-        "phpunit/phpunit": "^9.5"
+        "mikey179/vfsstream": "^1.6.11",
+        "phpunit/phpunit": "^9.6.13"
     },
     "autoload-dev": {
         "psr-4": {
@@ -66,11 +71,12 @@
     "extra": {
         "branch-alias": {
             "dev-main": "3.0-dev"
-        }
+        },
+        "class": "Composer\\Satis\\Plugin\\SatisPlugin"
     },
     "config": {
-        "platform": {
-            "php": "7.4.28"
+        "allow-plugins": {
+            "composer/satis": true
         }
     },
     "scripts": {
