@@ -20,9 +20,12 @@ class ArchiveBuilderHelper
 {
     /** @var OutputInterface The output Interface. */
     private $output;
-    /** @var array The 'archive' part of a configuration file. */
+    /** @var array<string, mixed> The 'archive' part of a configuration file. */
     private $archiveConfig;
 
+    /**
+     * @param array<string, mixed> $archiveConfig
+     */
     public function __construct(OutputInterface $output, array $archiveConfig)
     {
         $this->output = $output;
@@ -59,13 +62,13 @@ class ArchiveBuilderHelper
 
         $names = $package->getNames();
 
-        if ($this->archiveConfig['whitelist'] && !$this->isOneOfNamesInList($names, $this->archiveConfig['whitelist'])) {
+        if (0 !== count($this->archiveConfig['whitelist']) && !$this->isOneOfNamesInList($names, $this->archiveConfig['whitelist'])) {
             $this->output->writeln(sprintf("<info>Skipping '%s' (is not in whitelist)</info>", $name));
 
             return true;
         }
 
-        if ($this->archiveConfig['blacklist'] && $this->isOneOfNamesInList($names, $this->archiveConfig['blacklist'])) {
+        if (0 !== count($this->archiveConfig['blacklist']) && $this->isOneOfNamesInList($names, $this->archiveConfig['blacklist'])) {
             $this->output->writeln(sprintf("<info>Skipping '%s' (is in blacklist)</info>", $name));
 
             return true;
@@ -74,6 +77,10 @@ class ArchiveBuilderHelper
         return false;
     }
 
+    /**
+     * @param list<string> $names
+     * @param list<string> $list
+     */
     protected function isOneOfNamesInList(array $names, array $list): bool
     {
         $patterns = $this->convertListToRegexPatterns($list);
@@ -87,10 +94,13 @@ class ArchiveBuilderHelper
         return false;
     }
 
+    /**
+     * @param list<string> $patterns
+     */
     protected function doesNameMatchOneOfPatterns(string $name, array $patterns): bool
     {
         foreach ($patterns as $pattern) {
-            if (preg_match($pattern, $name)) {
+            if (1 === preg_match($pattern, $name)) {
                 return true;
             }
         }
@@ -98,6 +108,11 @@ class ArchiveBuilderHelper
         return false;
     }
 
+    /**
+     * @param list<string> $list
+     *
+     * @return list<string>
+     */
     protected function convertListToRegexPatterns(array $list): array
     {
         $patterns = [];
