@@ -22,5 +22,5 @@ interface BuilderInterface
      *
      * @param PackageInterface[] $packages List of packages to dump
      */
-    public function dump(array $packages);
+    public function dump(array $packages): void;
 }
