@@ -14,26 +14,39 @@ declare(strict_types=1);
 namespace Composer\Satis\Console;
 
 use Composer\Composer;
+use Composer\Console\Application as ComposerApplication;
 use Composer\Factory;
 use Composer\IO\ConsoleIO;
 use Composer\IO\IOInterface;
+use Composer\IO\NullIO;
 use Composer\Satis\Console\Command;
 use Composer\Satis\Satis;
 use Composer\Util\ErrorHandler;
+use Composer\Util\Platform;
 use Symfony\Component\Console\Application as BaseApplication;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class Application extends BaseApplication
+class Application extends ComposerApplication
 {
     /** @var IOInterface */
     protected $io;
-    /** @var Composer */
+    /** @var Composer|null */
     protected $composer;
 
     public function __construct()
     {
-        parent::__construct('Satis', Satis::VERSION);
+        // Composer constuctor does not allow passing BaseApplication::__construct('Satis', Satis::VERSION);
+        parent::__construct();
+        $this->setName('Satis');
+        $this->setVersion(Satis::VERSION);
+    }
+
+    /**
+     * Need to override composer's
+     */
+    public function __destruct()
+    {
     }
 
     public function doRun(InputInterface $input, OutputInterface $output): int
@@ -49,27 +62,9 @@ class Application extends BaseApplication
         return parent::doRun($input, $output);
     }
 
-    /**
-     * @param array|string|null $config
-     *  Either a configuration array or a filename to read from, if null it will read from the default filename
-     */
-    public function getComposer(bool $required = true, $config = null): Composer
-    {
-        if (null === $this->composer) {
-            try {
-                $this->composer = Factory::create($this->io, $config);
-            } catch (\InvalidArgumentException $e) {
-                $this->io->write($e->getMessage());
-                exit(1);
-            }
-        }
-
-        return $this->composer;
-    }
-
     protected function getDefaultCommands(): array
     {
-        $commands = array_merge(parent::getDefaultCommands(), [
+        $commands = array_merge(BaseApplication::getDefaultCommands(), [
             new Command\InitCommand(),
             new Command\AddCommand(),
             new Command\BuildCommand(),
@@ -77,5 +72,22 @@ class Application extends BaseApplication
         ]);
 
         return $commands;
+    }
+
+    public function getComposerWithConfig(mixed $config): ?Composer
+    {
+        if (null === $this->composer) {
+            try {
+                $this->composer = Factory::create(Platform::isInputCompletionProcess() ? new NullIO() : $this->io, $config, false, false);
+            } catch (\InvalidArgumentException $e) {
+                $this->io->writeError($e->getMessage());
+                if ($this->areExceptionsCaught()) {
+                    exit(1);
+                }
+                throw $e;
+            }
+        }
+
+        return $this->composer;
     }
 }
