@@ -29,8 +29,8 @@ class AddCommand extends BaseCommand
 {
     protected function configure(): void
     {
+        $this->getName() ?? $this->setName('add');
         $this
-            ->setName('add')
             ->setDescription('Add repository URL to satis JSON file')
             ->setDefinition([
                 new InputArgument('url', InputArgument::REQUIRED, 'VCS repository URL'),
@@ -57,7 +57,7 @@ class AddCommand extends BaseCommand
         $vcsDriver = $input->getOption('type');
         $repositoryName = $input->getOption('name');
 
-        if (preg_match('{^https?://}i', $configFile)) {
+        if (1 === preg_match('{^https?://}i', $configFile)) {
             $output->writeln('<error>Unable to write to remote file ' . $configFile . '</error>');
 
             return 2;
@@ -97,7 +97,7 @@ class AddCommand extends BaseCommand
 
         $repositoryConfig = ['type' => $vcsDriver, 'url' => $repositoryUrl];
 
-        if (!empty($repositoryName)) {
+        if (is_string($repositoryName) && '' !== $repositoryName) {
             $repositoryConfig['name'] = $repositoryName;
         }
 
@@ -122,12 +122,13 @@ class AddCommand extends BaseCommand
         $downloader = new HttpDownloader($io, $config);
         $repository = new VcsRepository(['url' => $repositoryUrl, 'type' => $type], $io, $config, $downloader);
 
-        if (!($driver = $repository->getDriver())) {
+        $driver = $repository->getDriver();
+        if (is_null($driver)) {
             return false;
         }
 
         $information = $driver->getComposerInformation($driver->getRootIdentifier());
 
-        return !empty($information['name']);
+        return isset($information['name']) && is_string($information['name']);
     }
 }
