@@ -1,5 +1,8 @@
 const Encore = require("@symfony/webpack-encore");
 const ESLintPlugin = require("eslint-webpack-plugin");
+const glob = require("glob");
+const path = require("path");
+const { PurgeCSSPlugin } = require("purgecss-webpack-plugin");
 
 Encore.addEntry("app", "./views/assets/js/app.js")
     .addStyleEntry("style", "./views/assets/css/style.scss")
@@ -9,7 +12,14 @@ Encore.addEntry("app", "./views/assets/js/app.js")
     .addPlugin(new ESLintPlugin())
     .enableSourceMaps(!Encore.isProduction())
     .setOutputPath("views/build/")
-    .setPublicPath("/build");
+    .setPublicPath("/build")
+    .addPlugin(
+        new PurgeCSSPlugin({
+            paths: glob.sync(`${path.join(__dirname, "views")}/**/*`, {
+                nodir: true,
+            }),
+        }),
+    );
 
 const config = Encore.getWebpackConfig();
 
